requirements.txt
django==5.1.4
djangorestframework==3.15.2
python-decouple==3.8
requests==2.32.3